# docs/swagger.yaml
definitions:
  github_com_dhnikolas_state-manager_internal_dto.Resource:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      created_at:
        type: string
      current_version:
        type: integer
      deletion_timestamp:
        type: string
      finalizers:
        items:
          type: string
        type: array
      kind:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
      resource_group:
        type: string
      shard_id:
        type: string
      spec:
        items:
          type: integer
        type: array
      status:
        items:
          type: integer
        type: array
      updated_at:
        type: string
      version:
        type: integer
    required:
    - shard_id
    type: object
  internal_http.CreateResourceRequest:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      finalizers:
        items:
          type: string
        type: array
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      shard_id:
        type: string
      spec:
        items:
          type: integer
        type: array
    required:
    - name
    - shard_id
    type: object
  internal_http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_http.UpdateResourceRequest:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      labels:
        additionalProperties:
          type: string
        type: object
      shard_id:
        type: string
      spec:
        items:
          type: integer
        type: array
    required:
    - shard_id
    type: object
  internal_http.UpdateResourceStatusRequest:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      current_version:
        type: integer
      labels:
        additionalProperties:
          type: string
        type: object
      shard_id:
        type: string
      status:
        items:
          type: integer
        type: array
      version:
        type: integer
    required:
    - shard_id
    type: object
info:
  contact: {}
paths:
  /api/v1/groups/{resource_group}/namespaces/{namespace}/kinds/{kind}/resources:
    post:
      consumes:
      - application/json
      description: Creates a new resource with the provided details.
      operationId: createResource
      parameters:
      - description: Resource Group
        in: path
        name: resource_group
        required: true
        type: string
      - description: Kind
        in: path
        name: kind
        required: true
        type: string
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Resource details
        in: body
        name: resource
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/internal_http.CreateResourceRequest'
          - properties:
              spec:
                additionalProperties: true
                type: object
            type: object
      produces:
      - application/json
      responses:
        "200":
          description: Resource created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_dhnikolas_state-manager_internal_dto.Resource'
            - properties:
                spec:
                  additionalProperties: true
                  type: object
                status:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 'Invalid input" example={"error":"Validation failed: shard_id
            is required"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
        "500":
          description: 'Server error" example={"error":"Failed to create resource:
            database error"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
      summary: Create a new resource
      tags:
      - resources
  /api/v1/groups/{resource_group}/namespaces/{namespace}/kinds/{kind}/resources/{name}:
    delete:
      consumes:
      - application/json
      description: Delete a resource by its resource_group, kind, namespace, and name.
      operationId: deleteResource
      parameters:
      - description: Resource Group
        in: path
        name: resource_group
        required: true
        type: string
      - description: Kind
        in: path
        name: kind
        required: true
        type: string
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 'Invalid input" example={"error":"Deletion failed: resource_group
            is required"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
        "404":
          description: 'Not found" example={"error":"Resource not found: no rows"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
      summary: Delete a resource
      tags:
      - resources
    get:
      consumes:
      - application/json
      description: Retrieves a resource by its resource_group, kind, namespace, and
        name.
      operationId: getResource
      parameters:
      - description: Resource Group
        in: path
        name: resource_group
        required: true
        type: string
      - description: Kind
        in: path
        name: kind
        required: true
        type: string
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Resource found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_dhnikolas_state-manager_internal_dto.Resource'
            - properties:
                spec:
                  additionalProperties: true
                  type: object
                status:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 'Invalid input" example={"error":"Validation failed: resource_group
            is required"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
        "404":
          description: 'Not found" example={"error":"Resource not found: no rows"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
      summary: Get a resource by key
      tags:
      - resources
    put:
      consumes:
      - application/json
      description: Updates a resource with the provided details.
      operationId: updateResource
      parameters:
      - description: Resource Group
        in: path
        name: resource_group
        required: true
        type: string
      - description: Kind
        in: path
        name: kind
        required: true
        type: string
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name
        in: path
        name: name
        required: true
        type: string
      - description: Resource details
        in: body
        name: resource
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/internal_http.UpdateResourceRequest'
          - properties:
              spec:
                additionalProperties: true
                type: object
            type: object
      produces:
      - application/json
      responses:
        "200":
          description: Resource updated
          schema:
            allOf:
            - $ref: '#/definitions/github_com_dhnikolas_state-manager_internal_dto.Resource'
            - properties:
                spec:
                  additionalProperties: true
                  type: object
                status:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 'Invalid input" example={"error":"Validation failed: shard_id
            is required"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
        "500":
          description: 'Server error" example={"error":"Failed to update resource:
            database error"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
      summary: Update a resource
      tags:
      - resources
  /api/v1/groups/{resource_group}/namespaces/{namespace}/kinds/{kind}/resources/{name}/status:
    put:
      consumes:
      - application/json
      description: Updates a resource with the provided details.
      operationId: updateResourceStatus
      parameters:
      - description: Resource Group
        in: path
        name: resource_group
        required: true
        type: string
      - description: Kind
        in: path
        name: kind
        required: true
        type: string
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name
        in: path
        name: name
        required: true
        type: string
      - description: Resource details
        in: body
        name: resource
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/internal_http.UpdateResourceStatusRequest'
          - properties:
              status:
                additionalProperties: true
                type: object
            type: object
      produces:
      - application/json
      responses:
        "200":
          description: Resource updated
          schema:
            allOf:
            - $ref: '#/definitions/github_com_dhnikolas_state-manager_internal_dto.Resource'
            - properties:
                spec:
                  additionalProperties: true
                  type: object
                status:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 'Invalid input" example={"error":"Validation failed: shard_id
            is required"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
        "409":
          description: 'Invalid input" example={"error":"Version conflict: resource
            version not match"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
        "500":
          description: 'Server error" example={"error":"Failed to update resource:
            database error"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
      summary: Update a resource
      tags:
      - resources
  /api/v1/resources:
    get:
      consumes:
      - application/json
      description: возвращает ресурсы по фильтру
      operationId: listResources
      parameters:
      - in: query
        name: kind
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: namespace
        type: string
      - in: query
        name: offset
        type: integer
      - in: query
        name: pending
        type: boolean
      - in: query
        name: resource_group
        type: string
      - in: query
        name: shard_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of pending resources
          schema:
            allOf:
            - $ref: '#/definitions/github_com_dhnikolas_state-manager_internal_dto.Resource'
            - properties:
                spec:
                  additionalProperties: true
                  type: object
                status:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Invalid input" example={"error":"shard_ids is required"}
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
        "500":
          description: 'Server error" example={"error":"Failed to list pending resources:
            database error"}'
          schema:
            $ref: '#/definitions/internal_http.ErrorResponse'
      summary: List pending resources
      tags:
      - resources
swagger: "2.0"