# pkg/stateclient/model_update_resource_request.go
/*
No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

API version: 1.0.0
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package openapi

import (
	"encoding/json"
	"bytes"
	"fmt"
)

// checks if the UpdateResourceRequest type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &UpdateResourceRequest{}

// UpdateResourceRequest struct for UpdateResourceRequest
type UpdateResourceRequest struct {
	Annotations *map[string]string `json:"annotations,omitempty"`
	Labels *map[string]string `json:"labels,omitempty"`
	ShardId string `json:"shard_id"`
	Spec map[string]interface{} `json:"spec,omitempty"`
}

type _UpdateResourceRequest UpdateResourceRequest

// NewUpdateResourceRequest instantiates a new UpdateResourceRequest object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewUpdateResourceRequest(shardId string) *UpdateResourceRequest {
	this := UpdateResourceRequest{}
	this.ShardId = shardId
	return &this
}

// NewUpdateResourceRequestWithDefaults instantiates a new UpdateResourceRequest object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewUpdateResourceRequestWithDefaults() *UpdateResourceRequest {
	this := UpdateResourceRequest{}
	return &this
}

// GetAnnotations returns the Annotations field value if set, zero value otherwise.
func (o *UpdateResourceRequest) GetAnnotations() map[string]string {
	if o == nil || IsNil(o.Annotations) {
		var ret map[string]string
		return ret
	}
	return *o.Annotations
}

// GetAnnotationsOk returns a tuple with the Annotations field value if set, nil otherwise
// and a boolean to check if the value has been set.
func (o *UpdateResourceRequest) GetAnnotationsOk() (*map[string]string, bool) {
	if o == nil || IsNil(o.Annotations) {
		return nil, false
	}
	return o.Annotations, true
}

// HasAnnotations returns a boolean if a field has been set.
func (o *UpdateResourceRequest) HasAnnotations() bool {
	if o != nil && !IsNil(o.Annotations) {
		return true
	}

	return false
}

// SetAnnotations gets a reference to the given map[string]string and assigns it to the Annotations field.
func (o *UpdateResourceRequest) SetAnnotations(v map[string]string) {
	o.Annotations = &v
}

// GetLabels returns the Labels field value if set, zero value otherwise.
func (o *UpdateResourceRequest) GetLabels() map[string]string {
	if o == nil || IsNil(o.Labels) {
		var ret map[string]string
		return ret
	}
	return *o.Labels
}

// GetLabelsOk returns a tuple with the Labels field value if set, nil otherwise
// and a boolean to check if the value has been set.
func (o *UpdateResourceRequest) GetLabelsOk() (*map[string]string, bool) {
	if o == nil || IsNil(o.Labels) {
		return nil, false
	}
	return o.Labels, true
}

// HasLabels returns a boolean if a field has been set.
func (o *UpdateResourceRequest) HasLabels() bool {
	if o != nil && !IsNil(o.Labels) {
		return true
	}

	return false
}

// SetLabels gets a reference to the given map[string]string and assigns it to the Labels field.
func (o *UpdateResourceRequest) SetLabels(v map[string]string) {
	o.Labels = &v
}

// GetShardId returns the ShardId field value
func (o *UpdateResourceRequest) GetShardId() string {
	if o == nil {
		var ret string
		return ret
	}

	return o.ShardId
}

// GetShardIdOk returns a tuple with the ShardId field value
// and a boolean to check if the value has been set.
func (o *UpdateResourceRequest) GetShardIdOk() (*string, bool) {
	if o == nil {
		return nil, false
	}
	return &o.ShardId, true
}

// SetShardId sets field value
func (o *UpdateResourceRequest) SetShardId(v string) {
	o.ShardId = v
}

// GetSpec returns the Spec field value if set, zero value otherwise.
func (o *UpdateResourceRequest) GetSpec() map[string]interface{} {
	if o == nil || IsNil(o.Spec) {
		var ret map[string]interface{}
		return ret
	}
	return o.Spec
}

// GetSpecOk returns a tuple with the Spec field value if set, nil otherwise
// and a boolean to check if the value has been set.
func (o *UpdateResourceRequest) GetSpecOk() (map[string]interface{}, bool) {
	if o == nil || IsNil(o.Spec) {
		return map[string]interface{}{}, false
	}
	return o.Spec, true
}

// HasSpec returns a boolean if a field has been set.
func (o *UpdateResourceRequest) HasSpec() bool {
	if o != nil && !IsNil(o.Spec) {
		return true
	}

	return false
}

// SetSpec gets a reference to the given map[string]interface{} and assigns it to the Spec field.
func (o *UpdateResourceRequest) SetSpec(v map[string]interface{}) {
	o.Spec = v
}

func (o UpdateResourceRequest) MarshalJSON() ([]byte, error) {
	toSerialize,err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o UpdateResourceRequest) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	if !IsNil(o.Annotations) {
		toSerialize["annotations"] = o.Annotations
	}
	if !IsNil(o.Labels) {
		toSerialize["labels"] = o.Labels
	}
	toSerialize["shard_id"] = o.ShardId
	if !IsNil(o.Spec) {
		toSerialize["spec"] = o.Spec
	}
	return toSerialize, nil
}

func (o *UpdateResourceRequest) UnmarshalJSON(data []byte) (err error) {
	// This validates that all required properties are included in the JSON object
	// by unmarshalling the object into a generic map with string keys and checking
	// that every required field exists as a key in the generic map.
	requiredProperties := []string{
		"shard_id",
	}

	allProperties := make(map[string]interface{})

	err = json.Unmarshal(data, &allProperties)

	if err != nil {
		return err;
	}

	for _, requiredProperty := range(requiredProperties) {
		if _, exists := allProperties[requiredProperty]; !exists {
			return fmt.Errorf("no value given for required property %v", requiredProperty)
		}
	}

	varUpdateResourceRequest := _UpdateResourceRequest{}

	decoder := json.NewDecoder(bytes.NewReader(data))
	decoder.DisallowUnknownFields()
	err = decoder.Decode(&varUpdateResourceRequest)

	if err != nil {
		return err
	}

	*o = UpdateResourceRequest(varUpdateResourceRequest)

	return err
}

type NullableUpdateResourceRequest struct {
	value *UpdateResourceRequest
	isSet bool
}

func (v NullableUpdateResourceRequest) Get() *UpdateResourceRequest {
	return v.value
}

func (v *NullableUpdateResourceRequest) Set(val *UpdateResourceRequest) {
	v.value = val
	v.isSet = true
}

func (v NullableUpdateResourceRequest) IsSet() bool {
	return v.isSet
}

func (v *NullableUpdateResourceRequest) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableUpdateResourceRequest(val *UpdateResourceRequest) *NullableUpdateResourceRequest {
	return &NullableUpdateResourceRequest{value: val, isSet: true}
}

func (v NullableUpdateResourceRequest) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableUpdateResourceRequest) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}